pyproject.toml
[project]
name = "trading-coach"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.136.1",
    "groq>=1.2.0",
    "httpx>=0.28.1",
    "numpy>=2.4.4",
    "pgvector>=0.4.2",
    "pydantic>=2.13.3",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]
